improve kfp op
