+++ mlrun/__init__.py
@@ -12,11 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__version__ = '0.2.0'
+__version__ = '0.2.5'
 
-from .run import get_or_create_ctx, new_function, mlrun_op, notebook_to_function
+from .run import get_or_create_ctx, new_function, code_to_function
 from .db import get_run_db
 from .model import RunTemplate, NewRun, RunObject
+from .kfpops import mlrun_op
 from .config import populate
